# Fixed and optimized Decimal.GetHashCode