This fixes #2944, tests are in https://github.com/dotnet/corefx/pull/30116 and the same change for CoreRT is in https://github.com/dotnet/corert/pull/5886.

### x64
|  Method |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
|  Native | 5.274 ns | 0.0284 ns | 0.0101 ns |   1.00 |
| Managed | 2.748 ns | 0.0114 ns | 0.0041 ns |   0.52 |


### x86
|  Method |      Mean |     Error |    StdDev | Scaled |
|-------- |----------:|----------:|----------:|-------:|
|  Native | 7.664 ns | 0.0662 ns | 0.0236 ns |   1.00 |
| Managed | 5.391 ns | 0.0154 ns | 0.0055 ns |   0.70 |