@kasper3 2018-06-04
@pentp, did you mean to supersede @slavanap's https://github.com/dotnet/coreclr/pull/18251?